requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
python-dotenv==1.0.0
openai==1.10.0
httpx==0.26.0
#pydantic==2.6.0
pydantic==2.7.4
google-generativeai==0.7.2
